Remove duplicate section
The string `You can install multiple instances of the connector for high availability and load balancing. Also, all connections are out-bound: from the connector to the Auth0 Server, so in general no changes to the firewall need to be applied.` was displayed twice in these docs and this change is to remove one of those duplicate sections.