Avoid uncaughtException in troubleshoot.js when res is undefined
When testing connectivity to Auth0, if the script is unable to reach the test endpoint, then `res` will be undefined, and the attempt to log `res.statusCode` results in an `uncaughtException`. In consequence, the troubleshooting won't be completed successfully.

This PR makes sure to check for `res` before attempting to log `res.statusCode`.