Unit test for LDAP_CA_PATH functionality.
At present, this test is effective when run standalone (npm test -- test/ldap.tests.js),
but not as part of the larger suite. The reason is that package.json specifies tests
should be run with NODE_TLS_REJECT_UNAUTHORIZED=0. Doing so with this test has the effect
of making the test meaningless (it passes even if the setting of LDAP_CA_PATH is
commented out). To help deal with this, ldap.tests.js explicitly sets NODE_TLS_REJECT_UNAUTHORIZED
to 1 in its setup method.

Unfortunately, it appears the NODE_TLS_REJECT_UNAUTHORIZED parameter is
read by node only once, meaning that if all tests are run as a suite, either this
test runs first, which presumably will break other tests, or those tests run first,
rendering this test ineffective (it will pass, whether or not behavior is correct).

It may be that the best option is to rewrite the other tests to use a mock ldap server as does
this one; not only can that approach obviate the need for setting NODE_TLS_REJECT_UNAUTHORIZED,
but it could also make the tests runnable without what is presumably a test ldap server
running inside of Auth0's private network. I leave that decision to the maintainers.