Add support for LDAP user profile attributes selection.
Via new configuration option: `LDAP_USER_BY_NAME_ATTRS`, a value to pass
through to the underlying LDAPjs library as the `attributes` option
for `Users.getByUserName`.

Current behaviour is maintained, select the default set of attributes,
which equates to a null value for the attribute.

To include extra/extended attributes in addition to the defaults, use
a list containing an asterisk (meaning defaults) and the additional
attributes:

[ "*", "att1", "att2" ]

To include ALL attributes, defaults and extended:

[ "*", "+" ]

To select only specific attribute(s), specify each in the list:

[ "att1", "att2" ]