Add support for LDAP user profile attributes selection.
Via new configuration option: `LDAP_USER_BY_NAME_ATTRS`, a value to pass
through to the underlying LDAPjs library as the `attributes` option
for `Users.getByUserName`.

Current behaviour is maintained, select the default set of attributes,
which equates to a null value for the option.

To include operational attributes in addition to the defaults, use
a list containing an asterisk (meaning defaults) and the operational
attributes: `[ "*", "att1", "att2" ]`

To include ALL attributes, defaults and operational: `[ "*", "+" ]`

More generally, to select only specific attribute(s) (regular or
operational), specify each in the list: `[ "att1", "att2" ]`